+++ lib/jade/symbol.rb
@@ -103,11 +103,11 @@ module Jade
     end
 
     def predeclared_interop_function(name)
-      InteropFunction[nil, name, [], nil, nil, [], nil, nil]
+      InteropFunction[nil, name, [], nil, nil, [], nil, nil, []]
     end
 
-    def interop_function(name, params, return_type, interop_module_name, constraints: [])
-      InteropFunction[nil, name, params, return_type, interop_module_name, constraints, nil, nil]
+    def interop_function(name, params, return_type, interop_module_name, constraints: [], capabilities: [])
+      InteropFunction[nil, name, params, return_type, interop_module_name, constraints, nil, nil, capabilities]
     end
 
     def type_application(constructor, args, span)